rediseño panel de juego
